requirements.txt
gunicorn==20.1.0
discord==1.7.3
python-dotenv==0.20.0